+++ src/Core/Auras/AuraCategoryIds.lua
@@ -1152,6 +1152,8 @@ addon.Core.AuraCategoryIds = {
 		[77606] = true, -- Dark Simulacrum
 		[236273] = true, -- Duel
 		[198819] = true, -- Sharpen Blade
+		[80240] = true, -- Havoc
+		[356723] = true, -- Scorpid Venom
 	},
 
 	-- Announced only when the player asks for them. These land often enough that a spoken name