+++ src/Core/Auras/AuraTtsSounds.lua
@@ -155,7 +155,9 @@ addon.Core.AuraTtsSounds = {
 		[360194] = "Deathmark.ogg", -- Deathmark
 		[236273] = "Duel.ogg", -- Duel
 		[274837] = "FeralFrenzy.ogg", -- Feral Frenzy
+		[80240] = "Havoc.ogg", -- Havoc
 		[385627] = "Kingsbane.ogg", -- Kingsbane
+		[356723] = "ScorpidVenom.ogg", -- Scorpid Venom
 		[198819] = "SharpenBlade.ogg", -- Sharpen Blade
 	},
 }